third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-console-api-repeated-in-console.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/inspector-protocol-test.js"></script>
<script>

function test()
{
    var expectedMessages = 4;
    var messages = [];

    InspectorTest.eventHandler["Runtime.consoleAPICalled"] = consoleAPICalled;
    InspectorTest.eventHandler["Console.messageAdded"] = messageAdded;
    InspectorTest.sendCommandOrDie("Runtime.enable", {});
    InspectorTest.sendCommandOrDie("Console.enable", {});
    InspectorTest.sendCommandOrDie("Runtime.evaluate", { "expression": "console.log(42)" });
    InspectorTest.sendCommandOrDie("Runtime.evaluate", { "expression": "console.error('abc')" });

    function consoleAPICalled(result)
    {
        messages.push("api call: " + result.params.args[0].value);
        if (!(--expectedMessages))
            done();
    }

    function messageAdded(result)
    {
        messages.push("console message: " + result.params.message.text);
        if (!(--expectedMessages))
            done();
    }

    function done()
    {
        messages.sort();
        for (var message of messages)
            InspectorTest.log(message);
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onLoad="runTest();">
Check that console.log is reported through Console domain as well.
</body>
</html>
