third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-console-api-repeated-in-console-expected.txt

CONSOLE MESSAGE: line 1: 42
CONSOLE ERROR: line 1: abc
Check that console.log is reported through Console domain as well.
api call: 42
api call: abc
console message: 42
console message: abc
